requirements.txt
opencv-python
pytesseract
Pillow
nltk
pandas
scikit-learn
tk
xml